fixed issue with useDimensions and applied change
