+++ src/components/blocks/_animators/IsentropicAnimator/IsentropicAnimator.module.scss
