+++ src/components/blocks/_animators/SoundingAnimator/SoundingAnimator.module.scss
