+++ src/components/blocks/_animators/SurfaceMapsAnimator/SurfaceMapsAnimator.tsx
@@ -9,7 +9,7 @@ import { faDownload, faInfoCircle, faLayerGroup, faWarning } from '@fortawesome/
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useParams } from 'next/navigation'
 import React, { useEffect, useState } from 'react'
-import ProductInfo, { ProductInfoProps } from '../ProductInfo/ProductInfo'
+import ProductInfo, { ProductInfoProps } from '../../ProductInfo/ProductInfo'
 import styles from './SurfaceMapsAnimator.module.scss'
 
 interface SurfaceMapsAnimatorProps {
@@ -20,7 +20,7 @@ const SurfaceMapsAnimator: React.FC<SurfaceMapsAnimatorProps> = ({ productInfo }
 	const { surfaceProductId: productId, surfaceRegionId: regionId, surfaceSiteId: siteId } = useParams()
 	const [activeTab, setActiveTab] = useState(-1)
 	const surfaceMapsNumberOfFrames = useRootStore.use.surfaceMapsNumberOfFrames()
-	const [wrapperRef, { width: width, height: height }] = useDimensions(8 / 6)
+	const [wrapperRef, { adjustedWidth, adjustedHeight }] = useDimensions(8 / 6, true)
 	const [surfaceMapsData, setSurfaceMapsData] = useState([])
 
 	useEffect(() => {
@@ -34,8 +34,8 @@ const SurfaceMapsAnimator: React.FC<SurfaceMapsAnimatorProps> = ({ productInfo }
 	return (
 		<div className={styles.surfaceMapsAnimatorContainer}>
 			<div className={styles.surfaceMapsAnimator} ref={wrapperRef}>
-				<div className={styles.animatorWrapper} style={{ width: width > height ? height : width, height: width > height ? height : width }}>
-					<Animator frames={surfaceMapsData} />
+				<div className={styles.animatorWrapper} style={{ width: adjustedWidth, height: adjustedHeight }}>
+					<Animator frames={surfaceMapsData} ratio={8 / 6} />
 				</div>
 			</div>
 			<Tabs activeTab={activeTab} setActiveTab={setActiveTab}>
